// HomeHub.Web/src/app/core/enums/announcements/additional-detail.enum.ts
export enum AdditionalDetailEnum {
    Internet = 1,
    CableTv = 2,
    Phone = 3,
    AntiBurglaryBlinds = 4,
    Monitoring = 5,
    AntiBurglaryDoors = 6,
    AntiBurglaryWindows = 7,
    AlarmSystem = 8,
    EntryPhone = 9,
    Videophone = 10,
    ClosedArea = 11,
    Furniture = 12,
    Fridge = 13,
    Tv = 14,
    WashingMachine = 15,
    Cooker = 16,
    Dishwasher = 17,
    Oven = 18,
    Balcony = 19,
    Basement = 20,
    Elevator = 21,
    AirConditioning = 22,
    UtilityRoom = 23,
    Garden = 24,
    TwoStory = 25,
    Garage = 26,
    ParkingSpace = 27,
    Terrace = 28,
    SeparateKitchen = 29
}